requirements.txt
Django==1.10.3
django-taggit==0.22.1
Pillow==4.2.1
Scrapy==1.4.0
